chore(mikro-orm): simplify testing setup
